Move tasks client to inside client module
