+++ platiagro/__init__.py
@@ -7,7 +7,6 @@ from platiagro.featuretypes import infer_featuretypes, validate_featuretypes, \
 from platiagro.figures import list_figures, save_figure
 from platiagro.metrics import list_metrics, save_metrics
 from platiagro.models import load_model, save_model
-from platiagro.tasks import create_task
 
 __all__ = ["client",
            "download_artifact",
@@ -17,7 +16,6 @@ __all__ = ["client",
            "DATETIME", "CATEGORICAL", "NUMERICAL",
            "list_figures", "save_figure",
            "list_metrics", "save_metrics",
-           "load_model", "save_model",
-           "create_task"]
+           "load_model", "save_model"]
 
 __version__ = "0.2.0"
