+++ platiagro/client/__init__.py
@@ -3,7 +3,9 @@ from platiagro.client.experiments import list_experiments, get_experiment_by_nam
 from platiagro.client.deployments import list_deployments, get_deployment_by_name, \
     run_deployment
 from platiagro.client.projects import list_projects, get_project_by_name
+from platiagro.client.tasks import create_task
 
 __all__ = ["list_experiments", "get_experiment_by_name", "run_experiment",
            "list_deployments", "get_deployment_by_name", "run_deployment",
-           "list_projects", "get_project_by_name"]
+           "list_projects", "get_project_by_name",
+           "create_task"]
